# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "HumStaffing application",
        "description": "HumStaffing application"
    },
    "basePath": "/",
  "tags": [
    {
      "name": "Auth",
      "description": "Auth API for authorize users in the system"
    }
   
  ],
  "securityDefinitions": {
    "Bearer" : {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },

    "paths": {
        "/signUp": {
            "post": {
                  "tags": ["Auth"],
        "summary": "Sign up new user in system",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SignInModel"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
            }
        },
         "/emailVrifyOtp": {
            "post": {
                  "tags": ["Auth"],
        "summary": "Email varification System Through OTP",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EmailVarify"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
            }
        },
         "/Login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login user in system",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "access_key": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
        },
          "/passwordchangeotpSend": {
            "post": {
                  "tags": ["Auth"],
        "summary": "Password Change System Through OTP",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ChangePassword"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
            }
        },
         "/changePassword": {
            "post": {
                  "tags": ["Auth"],
        "summary": "Type New password",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewPassword"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
            }
        },
         "/CreateProfile": {
                 "post": {
        "security": [{"Bearer": []}],
        "tags": ["CreateProfile"],
        "summary": "Create user collection",
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "in": "formData",
            "name": "profile",
            "type": "file",
            "required": true,
            "description": "Files to upload"
          },
           {
            "in": "formData",
            "name": "resume",
            "type": "file",
            "required": true,
            "description": "Files to upload"
          },
          
          {
            "in": "formData",
            "name": "firstname",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "lastname",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "category",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "gender",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "hourlyrange",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "experience",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "aboutme",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "qalification",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "certification",
            "type": "string"
          },
         
          {
            "in": "formData",
            "name": "speciality",
            "type": "string",
            "required": true
          },
           {
            "in": "formData",
            "name": "Location",
            "type": "string",
            "required": true
          },
           {
            "in": "formData",
            "name": "fburl",
            "type": "string",
            "required": true
          },
           {
            "in": "formData",
            "name": "instaurl",
            "type": "string",
            "required": true
          },
           {
            "in": "formData",
            "name": "linkedinurl",
            "type": "string",
            "required": true
          },
           {
            "in": "formData",
            "name": "twitterurl",
            "type": "string",
            "required": true
          }
          
          
          
          
          
        ],
        "responses": {
          "201": {
            "description": "OK",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/CreateProfile"
            }
          }
        }
      }
        },
         "/BasicInfo/:id": {
      "patch": {
        "security": [{"Bearer": []}],
        "tags": ["BasicInfo"],
        "summary": "/BasicInfo/:id",
        "parameters": [
          {
            "name": "profile_ID ",
            "in": "path",

            "type": "string",
            "required": true,
            "description": "Claim id of one of claims"
           
          
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                
                "firstname": {
                  "type": "string"
                },
                "lastname": {
                  "type": "string"
                },
                "gender": {
                  "type": "string"
                },
                "experience": {
                  "type": "string"
                },
                "hourlyrange": {
                  "type": "string"
                },
                 "city": {
                  "type": "string"
                }

              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/BasicInfo"
            }
          }
        }
      }
    },
     "/addJob": {
            "post": {
                  "tags": ["Auth"],
        "summary": "Add Jobs Assignment System",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/addJob"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
            }
        }
        


    },
    "definitions": {
   
    "SignInModel": {
      "properties": {
       
       "fullname":{
        "type":"string"
       },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
       
       
       
      }
    },
    "EmailVarify": {
      "properties": {
       
      
        "email": {
          "type": "string"
        },
        "code": {
          "type": "string"
        }
       
       
       
      }
    },
     "LoginModel": {
      "required": ["email", "password"],
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
     "ChangePassword": {
      "required": ["email"],
      "properties": {
        "email": {
          "type": "string"
        }
       

      }
    },
   "NewPassword":{
  "properties": {
       
      
        "email": {
          "type": "string"
        },
        "code": {
          "type": "string"
        },
       "password":{
        "type":"string"
       }
       
       
      }
   },
   "CreateProfile":{
  "properties": {
       
      
        
     "firstname":{
      "type":"string"
     },
     "lastname":{

      "type":"string"
     },
     "category":{
      "type":"string"
     },
     "gender":{
      "type":"string"
     },
     "hourlyrange":{
      "type":"string"
     },
     "experience":{
      "type":"string"
     },
     "aboutme":{
      "type":"string"
     },
     "qalification":{
      "type":"string"
     },
     "certification":{
      "type":"string"
     },
     "speciality":{
      "type":"string"
     }


       
      }
   },
      "BasicInfo": {
      "type": "object",
      "properties": {
        
        "firstname": {
          "type": "string"
        },
        "lastname": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "hourlyrange": {
          "type": "string"
        },
        "experience": {
          "type": "string"
        }
      }
    },
     "addJob": {
      "type": "object",
      "properties": {
        
        "positionTitle": {
          "type": "string"
        },
        "hospital_Faculty": {
          "type": "string"
        },
        "speciality": {
          "type": "string"
        },
        "jod_duration": {
          "type": "string"
        },
        "hourlyRate": {
          "type": "integer"
        },
        "shift": {
          "type": "string"
        },
        "from": {
          "type": "integer"
        },
        "to": {
          "type": "integer"
        },
        "startDate": {
          "type": "string"
        },
        "endDate": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
       

      }
    }
  



  }

    
}